+++ packages/ui/src/blocks/inputs-debug/visually-hidden-input.tsx
@@ -159,4 +159,3 @@ function VisuallyHiddenInput<T = InputValue>(
 }
 
 export { VisuallyHiddenInput }
-export type { VisuallyHiddenInputProps }